+++ stroom-security/stroom-security-client/src/main/java/stroom/security/client/presenter/AbstractDataUserListPresenter.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2016 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package stroom.security.client.presenter;
 
 import stroom.dispatch.client.RestFactory;
@@ -23,16 +7,16 @@ import stroom.widget.popup.client.event.HidePopupEvent;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 
-public class AdvancedUserListPresenter extends AbstractUserListPresenter {
+public abstract class AbstractDataUserListPresenter extends AbstractUserListPresenter {
 
     private final RestFactory restFactory;
     private UserDataProvider dataProvider;
     private FindUserCriteria findUserCriteria;
 
     @Inject
-    public AdvancedUserListPresenter(final EventBus eventBus,
-                                     final UserListView userListView,
-                                     final RestFactory restFactory) {
+    public AbstractDataUserListPresenter(final EventBus eventBus,
+                                         final UserListView userListView,
+                                         final RestFactory restFactory) {
         super(eventBus, userListView);
         this.restFactory = restFactory;
     }
@@ -43,14 +27,18 @@ public class AdvancedUserListPresenter extends AbstractUserListPresenter {
         registerHandler(getSelectionModel().addSelectionHandler(event -> {
             if (event.getSelectionType().isDoubleSelect()) {
                 if (findUserCriteria != null && findUserCriteria.getRelatedUser() == null) {
-                    HidePopupEvent.fire(
-                            AdvancedUserListPresenter.this,
-                            AdvancedUserListPresenter.this,
-                            false,
-                            true);
+                    hide();
                 }
             }
         }));
+    }
+
+    void hide() {
+        HidePopupEvent.fire(
+                this,
+                this,
+                false,
+                true);
     }
 
     @Override