+++ stroom-security/stroom-security-impl/src/main/java/stroom/security/impl/UserNameResourceImpl.java
@@ -1,0 +1,55 @@
+/*
+ *
+ *   Copyright 2017 Crown Copyright
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package stroom.security.impl;
+
+import stroom.event.logging.rs.api.AutoLogged;
+import stroom.event.logging.rs.api.AutoLogged.OperationType;
+import stroom.security.shared.FindUserNameCriteria;
+import stroom.security.shared.UserNameProvider;
+import stroom.security.shared.UserNameResource;
+import stroom.util.shared.ResultPage;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import javax.inject.Inject;
+
+@AutoLogged(OperationType.MANUALLY_LOGGED)
+class UserNameResourceImpl implements UserNameResource {
+
+    private final Set<UserNameProvider> userNameProviders;
+
+    @Inject
+    public UserNameResourceImpl(final Set<UserNameProvider> userNameProviders) {
+        this.userNameProviders = userNameProviders;
+    }
+
+    @Override
+    public ResultPage<String> find(final FindUserNameCriteria criteria) {
+        final Set<String> names = new HashSet<>();
+        for (final UserNameProvider userNameProvider : userNameProviders) {
+            final ResultPage<String> resultPage = userNameProvider.findUserNames(criteria);
+            names.addAll(resultPage.getValues());
+        }
+        final List<String> list = names.stream().sorted().collect(Collectors.toList());
+        return new ResultPage<>(list);
+    }
+}
+